move base64 buffer serialization to sidecar
