+++ crates/sidecar/Cargo.toml
@@ -4,6 +4,8 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+base64 = "0.22.0"
+bytes = "1.5.0"
 clap = { version = "4.5.1", features = ["derive"] }
 serde = { version = "1.0.196", features = ["derive"] }
 serde_json = "1.0.113"
